requirements.txt
moviepy
google.generativeai
fastapi
os
time
ffmpeg